Core: Add api_v0 application
This allows a better separation of concerns: api_v0 should be where the
code for the API frontend goes.

We add api_v0 to the settings.py and urls.py so that the URLs are
properly nested in /api/0/ (as there may be another API version at
/api/1/ at some point). This allows for much easier deprecation and
innovation in the API endpoints whilst keeping compatibility with old
applications for a while.