Upgrade to Django 1.10
Configure the new LOGGING settings so that the behaviour is still
similar to Django 1.9 and errors come out through the console.

Use newer Argon2 password hasher by default, and require it in
requirements.txt

Argon2 (https://en.wikipedia.org/wiki/Argon2) won the Password Hashing
Competition in July 2015, so it seems stronger than the default PBKDF2
and is recommended as per Django guidelines (it's not enabled by default
because it requires an external dependency).

MIDDLEWARE_CLASSES is updated to MIDDLEWARE because we are already using
a corsheaders module supporting the newer Django 1.10 MIDDLEWARE.