Remove the failing accounts test.
The test had several intrinsic problems, but most importantly, it did
not test what it was supposed to. This escaped code-review.

The test was created to test whether a user could still access the API
after changing their password. However, as no real authentication
mechanism is provided, but instead the user is forced onto an
authenticated request, via `.force_login()`, it is *impossible* to make
the requests presently invalid.

If such behavior is desired, it should be implemented and tested in the
proper authentication backend, most probably in our own jwt_knox
application, where we have access to tokens and we don't really need
cross-application coupling.