Add Travis configuration file (#9)
Add a minimal Travis configuration file to enable testing on Travis-CI
for continuous integration so that we have up-to-date test and
regressions information.

* Allow Travis failures on some Python versions

We are not currently supporting a wide variety of Python versions, so we
will allow failures on those "officially unsupported" versions.

* Turn on cache for pip packages in Travis

This should remove some of the lag caused by building the many packages
which require C extensions, such as cryptography or argon2.

* Add latest pypy version support to Travis

Travis does not build PyPy > 2.6 (needed for cryptography), and thus
fails for no reason before even testing our application.

* Always check package upgrades in Travis

Travis is instructed to cache pip results. We wtill tell pip to
look for package upgrades and possibly update the cache.